Rotten Tomatoes
We have a movie rating site with two critics who have different, structured, well respected tastes. The critics both watch movies and then rate them out of 10. We want to explore the ratings to see if there is any patterns which might shine light on good ways to categorize movies. The critics have rated 5 movies and given the ratings: (1, 4), (5, 6), (7, 5), (7, 8), (2, 4) such that the first critics rating is the first element in the tuple and the second critics rating is the second.
Walk through the k-means algorithm starting with your centroids at (6, 7) and (7, 8). Stop when your centroids no longer change. What are the final two centroids? What are the final assignments of points to clusters?
Use hierarchical agglomerative clustering